+++ CHANGELOG.md
@@ -13,6 +13,9 @@ tech changes will usually be stripped from release notes for the public
 ### Added
 
 -   New quick toggle to disable LoS rendering for the DM only
+-   Forgot password flow if Mail is configured on the server
+    -   This only works if the user account actually has an email-address
+-   [server] Email configuration setup
 
 ### Changed